fix(chat): fix footer displaying logic (Issue #3465)
